so modified condor script generator to use correct naming convention (had to add more lines from this magical csv stuff to make it work correctly).
modified runExp so that usage exists and gives out a message saying whats
needed. also adjusted output-dir as it currently checked for outputdir in one
place, and output-dir in another area. Also added elif instead of ifs in parsde
args as it was breaking out

modified sneemediator so that the avrora script command does not have java in
it.